Partial category group APIs

+++ src/services/account.service.ts
@@ -1,13 +1,19 @@
 import { ERROR_MESSAGE } from '../common/constants';
-import { account } from '../db/schema';
+import { account } from '@/db/schemas';
 import { InternalServerError } from '../exception/exception';
 import { db } from '../providers/db-config';
 import { eq } from 'drizzle-orm';
-import { fromPromise } from 'neverthrow';
+import { fromPromise, ok, ResultAsync } from 'neverthrow';
 import { safeDbFetch } from '../utils/safe-db-fetch';
+import { Account } from '../db/db.types';
 
 export class AccountService {
-  getById = (id: number) => {
+  /**
+   * Retrieves an account by its ID
+   * @param {number} id - The ID of the account to retrieve
+   * @returns {ResultAsync<Account, Error>} A Result containing the account or an error
+   */
+  getById = (id: number): ResultAsync<Account, Error> => {
     const query = db.query.account.findFirst({
       where: eq(account.id, id),
     });
@@ -15,8 +21,11 @@ export class AccountService {
     const result = safeDbFetch(query, 'account');
     return result;
   };
-
-  list = () => {
+  /**
+   * Lists all accounts
+   * @returns {Result<Account[], InternalServerError>} A Result containing an array of accounts or an error
+   */
+  list = (): ResultAsync<Account[], InternalServerError> => {
     const promise = db.query.account.findMany();
     const result = fromPromise(promise, () => {
       return new InternalServerError({
@@ -30,7 +39,12 @@ export class AccountService {
     });
     return result;
   };
-
+  /**
+   * Creates a new account
+   * @param {Object} payload - The account data
+   * @param {string} payload.name - The name of the account
+   * @returns {ResultAsync<ResultSet, Error>} A Result containing the created account or an error
+   */
   create = (payload: { name: string }) => {
     const query = fromPromise(
       db.insert(account).values(payload).returning(),
@@ -48,9 +62,16 @@ export class AccountService {
     return query;
   };
 
+  /**
+   * Updates an account by its ID
+   * @param {number} id - The ID of the account to update
+   * @param {Object} payload - The account data to update
+   * @param {string} payload.name - The new name of the account
+   * @returns {ResultAsync<ResultSet, Error>} A Result containing the update result or an error
+   */
   update = (id: number, payload: { name: string }) => {
     const query = fromPromise(
-      db.update(account).set(payload).where(eq(account.id, id)).returning(),
+      db.update(account).set(payload).where(eq(account.id, id)),
       () => {
         return new InternalServerError({
           message: ERROR_MESSAGE.serverError,
@@ -65,21 +86,36 @@ export class AccountService {
     return query;
   };
 
-  delete = (id: number) => {
-    const query = fromPromise(
-      db.delete(account).where(eq(account.id, id)),
-      () => {
-        return new InternalServerError({
-          message: ERROR_MESSAGE.serverError,
-          details: {
-            target: 'account',
-            operation: 'delete',
-            layer: 'db',
-          },
-        });
+  delete = (id: number): ResultAsync<number, InternalServerError> => {
+    const internalServerError = new InternalServerError({
+      message: ERROR_MESSAGE.serverError,
+      details: {
+        target: 'account',
+        operation: 'delete',
+        layer: 'db',
+      },
+    });
+    return fromPromise(
+      db.query.account.findFirst({ where: eq(account.id, id) }),
+      () => internalServerError
+    ).andThen((accountRecord) => {
+      if (!accountRecord) {
+        return ok(404); // Gone: Account not found
       }
-    );
-    return query;
+
+      return fromPromise(
+        db.delete(account).where(eq(account.id, id)),
+        () =>
+          new InternalServerError({
+            message: ERROR_MESSAGE.serverError,
+            details: {
+              target: 'account',
+              operation: 'delete',
+              layer: 'db',
+            },
+          })
+      ).map(() => 204); // No Content: Account successfully deleted
+    });
   };
 }
